requirements.txt
torch
transformers
datasets
accelerate
scikit-learn
rouge_score
evaluate
uvicorn
fastapi
python-dotenv
peft